+++ src/marionette.eventaggregator.js
@@ -1,8 +1,0 @@
-// Event Aggregator
-// ----------------
-// A pub-sub object that can be used to decouple various parts
-// of an application through event-driven architecture.
-//
-// Extends [Backbone.Wreqr.EventAggregator](https://github.com/marionettejs/backbone.wreqr)
-// and mixes in an EventBinder from [Backbone.EventBinder](https://github.com/marionettejs/backbone.eventbinder).
-Marionette.EventAggregator = Backbone.Wreqr.EventAggregator;